ft_transcendence/backend/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
cryptography==43.0.0
daphne==4.1.2
distlib==0.3.8
distro==1.9.0
Django==4.2.13
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.2
Faker==26.0.0
filelock==3.15.4
grpcio==1.65.5
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
jiter==0.5.0
msgpack==1.0.8
openai==1.41.0
pillow==10.4.0
platformdirs==4.2.2
proto-plus==1.24.0
protobuf==4.25.4
psycopg-binary==3.1.19
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
pyOpenSSL==24.2.1
pyotp==2.9.0
pyparsing==3.1.2
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
qrcode==7.4.2
redis==5.0.8
requests==2.32.3
rsa==4.9
service-identity==24.1.0
shortuuid==1.0.13
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
tqdm==4.66.5
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
virtualenv==20.26.3
websockets==12.0
zope.interface==6.4.post2
whitenoise==6.7.0